requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
django-dotenv==1.4.2
djangorestframework==3.16.1
python-dotenv==1.2.1
sqlparse==0.5.3
